MySQL と SQLite3 の共通部分を集約した
