+++ src/server/Model/DB/DBBase.ts
@@ -1,20 +1,32 @@
 import Model from '../Model';
+import DBOperator from './DBOperator';
 
 /**
 * DBBase クラス
 */
 abstract class DBBase extends Model {
+    protected operator: DBOperator;
+
+    constructor(operator: DBOperator) {
+        super();
+        this.operator = operator;
+    }
+
     /**
     * ping
     * @return Promise<void>
     */
-    abstract ping(): Promise<void>;
+    public ping(): Promise<void> {
+        return this.operator.ping();
+    }
 
     /**
     * end
     * @return Promise<void>
     */
-    abstract end(): Promise<void>;
+    public end(): Promise<void> {
+        return this.operator.end();
+    }
 }
 
 export default DBBase;
