+++ src/server/Model/DB/MySQL/MySQLRecordedDB.ts
@@ -1,0 +1,57 @@
+import * as DBSchema from '../DBSchema';
+import { RecordedDB } from '../RecordedDB';
+
+class MySQLRecordedDB extends RecordedDB {
+    /**
+    * create table
+    * @return Promise<void>
+    */
+    public create(): Promise<void> {
+        let query = `CREATE TABLE IF NOT EXISTS ${ DBSchema.TableName.Recorded } (`
+            + 'id int primary key auto_increment, '
+            + 'programId bigint not null, '
+            + 'channelId bigint not null, '
+            + 'channelType text not null, '
+            + 'startAt bigint not null, '
+            + 'endAt bigint not null, '
+            + 'duration bigint not null, '
+            + 'name text not null, '
+            + 'description text null, '
+            + 'extended text null, '
+            + 'genre1 integer null, '
+            + 'genre2 integer null, '
+            + 'videoType text null, '
+            + 'videoResolution text null, '
+            + 'videoStreamContent integer null, '
+            + 'videoComponentType integer null, '
+            + 'audioSamplingRate integer null, '
+            + 'audioComponentType integer null, '
+            + 'recPath text, '
+            + 'ruleId int, '
+            + 'thumbnailPath text, '
+            + 'recording boolean '
+        + ') engine=InnoDB;'
+
+        return this.operator.runQuery(query);
+    }
+
+    /**
+    * recording 状態を解除する
+    * @param id: recorded id
+    * @return Promise<void>
+    */
+    public removeRecording(id: number): Promise<void> {
+        return this.operator.runQuery(`update ${ DBSchema.TableName.Recorded } set recording = false where id = ${ id }`);
+    }
+
+    /**
+    * recording 状態をすべて解除する
+    * @return Promise<void>
+    */
+    public removeAllRecording(): Promise<void> {
+        return this.operator.runQuery(`update ${ DBSchema.TableName.Recorded } set recording = false where recording = ${ true }`);
+    }
+}
+
+export default MySQLRecordedDB;
+
