--- src/server/Model/DB/SQLite3/EncodedDB.ts
@@ -1,100 +1,0 @@
-import * as path from 'path';
-import SQLite3Base from './SQLite3Base';
-import * as DBSchema from '../DBSchema';
-import Util from '../../../Util/Util';
-import { EncodedDBInterface } from '../EncodedDB';
-
-class EncodedDB extends SQLite3Base implements EncodedDBInterface {
-    /**
-    * create table
-    * @return Promise<void>
-    */
-    public create(): Promise<void> {
-        let query = `create table if not exists ${ DBSchema.TableName.Encoded } (`
-            + 'id integer primary key autoincrement, '
-            + 'recordedId integer, '
-            + 'name text not null, '
-            + 'path text not null, '
-            + `foreign key(recordedId) references ${ DBSchema.TableName.Recorded }(id) `
-        + ');'
-
-        return this.runQuery(query);
-    }
-
-    /**
-    * encoded 挿入
-    * @param recordedId: recordedId
-    * @param name: name
-    * @para, filePath: file path
-    * @return Promise<number> insertId
-    */
-    public insert(recordedId: number, name: string, filePath: string): Promise<number> {
-        let query = `insert into ${ DBSchema.TableName.Encoded } (`
-            + 'recordedId,'
-            + 'name, '
-            + 'path '
-        + ') VALUES ('
-            + '?, ?, ?'
-        + ');'
-
-        let value: any[] = [
-            recordedId,
-            name,
-            filePath.slice(Util.getRecordedPath().length + path.sep.length),
-        ];
-
-        return this.runInsert(query, value);
-    }
-
-    /**
-    * encoded を削除
-    * @param id: encoded id
-    * @return Promise<void>
-    */
-    public delete(id: number): Promise<void> {
-        return this.runQuery(`delete from ${ DBSchema.TableName.Encoded } where id = ${ id }`);
-    }
-
-    /**
-    * encoded を recorded id を指定して削除
-    * @param recordedId: recorded id
-    * @return Promise<void>
-    */
-    public deleteRecordedId(recordedId: number): Promise<void> {
-        return this.runQuery(`delete from ${ DBSchema.TableName.Encoded } where recordedId = ${ recordedId }`);
-    }
-
-    /**
-    * id 検索
-    * @param id: encoded id
-    * @return Promise<DBSchema.EncodedSchema | null>
-    */
-    public async findId(id: number): Promise<DBSchema.EncodedSchema | null> {
-        let programs = await this.runQuery(`select * from ${ DBSchema.TableName.Encoded } where id = ${ id }`);
-        return this.getFirst(this.fixResult(<DBSchema.EncodedSchema[]>programs));
-    }
-
-    /**
-    * @param programs: DBSchema.RecordedSchema[]
-    */
-    private fixResult(programs: DBSchema.EncodedSchema[]): DBSchema.EncodedSchema[] {
-        let baseDir = Util.getRecordedPath();
-        return programs.map((program) => {
-            // path をフルパスへ書き換える
-            program.path = path.join(baseDir, program.path);
-            return program;
-        });
-    }
-
-    /**
-    * recorded id 検索
-    * @param recordedId: recorded id
-    * @return Promise<DBSchema.EncodedSchema[]>
-    */
-    public async findRecordedId(recordedId: number): Promise<DBSchema.EncodedSchema[]> {
-        let programs = await this.runQuery(`select * from ${ DBSchema.TableName.Encoded } where recordedId = ${ recordedId }`);
-        return this.fixResult(<DBSchema.EncodedSchema[]>programs);
-    }
-}
-
-export default EncodedDB;
