+++ src/server/Model/DB/SQLite3/SQLite3EncodedDB.ts
@@ -1,0 +1,23 @@
+import * as DBSchema from '../DBSchema';
+import { EncodedDB } from '../EncodedDB';
+
+class SQLite3EncodedDB extends EncodedDB {
+    /**
+    * create table
+    * @return Promise<void>
+    */
+    public create(): Promise<void> {
+        let query = `create table if not exists ${ DBSchema.TableName.Encoded } (`
+            + 'id integer primary key autoincrement, '
+            + 'recordedId integer, '
+            + 'name text not null, '
+            + 'path text not null, '
+            + `foreign key(recordedId) references ${ DBSchema.TableName.Recorded }(id) `
+        + ');'
+
+        return this.operator.runQuery(query);
+    }
+}
+
+export default SQLite3EncodedDB;
+
