+++ src/server/Model/ServiceModelFactorySetting.ts
@@ -1,14 +1,16 @@
 import factory from './ModelFactory';
-import MySQLServicesDB from './DB/MySQL/ServicesDB';
-import MySQLProgramsDB from './DB/MySQL/ProgramsDB';
-import MySQLRulesDB from './DB/MySQL/RulesDB';
-import MySQLRecordedDB from './DB/MySQL/RecordedDB';
-import MySQLEncodedDB from './DB/MySQL/EncodedDB';
-import SQLite3ServicesDB from './DB/SQLite3/ServicesDB';
-import SQLite3ProgramsDB from './DB/SQLite3/ProgramsDB';
-import SQLite3RulesDB from './DB/SQLite3/RulesDB';
-import SQLite3RecordedDB from './DB/SQLite3/RecordedDB';
-import SQLite3EncodedDB from './DB/SQLite3/EncodedDB';
+import MySQLOperator from './DB/MySQL/MySQLOperator';
+import SQLite3Operator from './DB/SQLite3/SQLite3Operator';
+import MySQLServicesDB from './DB/MySQL/MySQLServicesDB';
+import MySQLProgramsDB from './DB/MySQL/MySQLProgramsDB';
+import MySQLRulesDB from './DB/MySQL/MySQLRulesDB';
+import MySQLRecordedDB from './DB/MySQL/MySQLRecordedDB';
+import MySQLEncodedDB from './DB/MySQL/MySQLEncodedDB';
+import SQLite3ServicesDB from './DB/SQLite3/SQLite3ServicesDB';
+import SQLite3ProgramsDB from './DB/SQLite3/SQLite3ProgramsDB';
+import SQLite3RulesDB from './DB/SQLite3/SQLite3RulesDB';
+import SQLite3RecordedDB from './DB/SQLite3/SQLite3RecordedDB';
+import SQLite3EncodedDB from './DB/SQLite3/SQLite3EncodedDB';
 import { RulesModel } from './Api/RulesModel';
 import { RecordedModel } from './Api/RecordedModel';
 import { ChannelsModel } from './Api/ChannelsModel';
@@ -37,11 +39,12 @@ namespace ModelFactorySetting {
     */
     export const init = (): void => {
         const isMysql = Util.getDBType() === 'mysql';
-        const servicesDB = isMysql ? new MySQLServicesDB() : new SQLite3ServicesDB();
-        const programsDB = isMysql ? new MySQLProgramsDB() : new SQLite3ProgramsDB();
-        const rulesDB = isMysql ? new MySQLRulesDB() : new SQLite3RulesDB();
-        const recordedDB = isMysql ? new MySQLRecordedDB() : new SQLite3RecordedDB();
-        const encodedDB = isMysql ? new MySQLEncodedDB() : new SQLite3EncodedDB();
+        const operator = isMysql ? new MySQLOperator() : new SQLite3Operator();
+        const servicesDB = isMysql ? new MySQLServicesDB(operator) : new SQLite3ServicesDB(operator);
+        const programsDB = isMysql ? new MySQLProgramsDB(operator) : new SQLite3ProgramsDB(operator);
+        const rulesDB = isMysql ? new MySQLRulesDB(operator) : new SQLite3RulesDB(operator);
+        const recordedDB = isMysql ? new MySQLRecordedDB(operator) : new SQLite3RecordedDB(operator);
+        const encodedDB = isMysql ? new MySQLEncodedDB(operator) : new SQLite3EncodedDB(operator);
 
         let encodeProcessManager = EncodeProcessManager.getInstance();
         EncodeManager.init(encodeProcessManager);
